fix: everything to work with paths with spaces

+++ src/fs/delete.js
@@ -2,6 +2,11 @@ import fs from "fs/promises";
 import path from "path";
 
 async function deleteFile(currentDir, fileName) {
+  if (!fileName) {
+    console.log("Error: Not enough arguments for 'create' command");
+    return;
+  }
+
   try {
     await fs.rm(path.resolve(currentDir, fileName));
     console.log(`File "${fileName}" deleted\n`);
